Validacion Susana Marta 100

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -14,8 +14,8 @@ class CreateUsersTable extends Migration
     public function up()
     {
         Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('role_id')->default('2');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('role_id')->default('2');
             $table->foreign('role_id')->references('id')->on('roles');
             $table->string('name');
             $table->string('lastname')->nullable();
