+++ database/migrations/2018_11_04_121357_create_condicion_facturacions_table.php
@@ -14,12 +14,12 @@ class CreateCondicionFacturacionsTable extends Migration
     public function up()
     {
         Schema::create('condicion_facturacions', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('empresa_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('empresa_id');
             $table->foreign('empresa_id')->references('id')->on('empresas')->onDelete('cascade');
-            $table->unsignedInteger('formapago_id');
+            $table->unsignedBigInteger('formapago_id');
             $table->foreign('formapago_id')->references('id')->on('forma_pagos');
-            $table->unsignedInteger('periodopago_id');
+            $table->unsignedBigInteger('periodopago_id');
             $table->foreign('periodopago_id')->references('id')->on('periodo_pagos');
             $table->integer('diafactura')->nullable();
             $table->integer('diavencimiento')->nullable();
