+++ database/migrations/2018_11_28_170537_create_table_user_empresa.php
@@ -14,10 +14,10 @@ class CreateTableUserEmpresa extends Migration
     public function up()
     {
         Schema::create('user_empresas', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->unsignedInteger('empresa_id');
+            $table->unsignedBigInteger('empresa_id');
             $table->foreign('empresa_id')->references('id')->on('empresas');
             $table->unique(['user_id', 'empresa_id']);
             $table->timestamps();
